fix: 快速单测按目标分进程执行 + terminal_bridge 编译与格式修复
- ci.yml：PR 快速路径此前把多个后端服务的测试目标放进同一个 pytest 进程。
  各服务顶层包名同为 app，且存在同名测试模块（api-gateway 与
  conversation-service 均有 tests/unit/test_bridge_logs.py），合并执行时
  先收集的服务占用 sys.modules['app']，导致 import file mismatch、
  Table already defined、ModuleNotFoundError 等与变更无关的伪失败。
  改为逐个目标独立进程执行（与完整路径既有的分进程策略一致），任一失败即退出。
- terminal_bridge：修复 markDroppedLogEmitted 的 int64/Duration 类型不匹配
  编译错误；修正测试中 LogHub.subs 字段类型；统一 gofmt 格式。

[env:staging:aihci][agent:codebuddy]

+++ terminal_bridge/main.go
@@ -2511,7 +2511,7 @@ var droppedLogEmittedAt int64
 func markDroppedLogEmitted() bool {
 	now := time.Now().UnixNano()
 	last := atomic.LoadInt64(&droppedLogEmittedAt)
-	if last != 0 && now-last < 5*time.Second {
+	if last != 0 && now-last < int64(5*time.Second) {
 		return false
 	}
 	return atomic.CompareAndSwapInt64(&droppedLogEmittedAt, last, now)
@@ -4070,18 +4070,18 @@ func main() {
 	// 否则排障时无法回答"日志到底写在哪"（工单 Q2026092010235 的直接教训）。
 	logHub.applyRuntimeMode(config.Mode)
 	blog("INFO", "bridge.startup", "terminal_bridge 已启动", "", "", "", "", map[string]any{
-		"version":        getVersion(),
-		"commit":         CommitID,
-		"built":          BuildTime,
-		"mode":           config.Mode,
-		"listen":         config.address(),
+		"version":         getVersion(),
+		"commit":          CommitID,
+		"built":           BuildTime,
+		"mode":            config.Mode,
+		"listen":          config.address(),
 		"allowed_origins": config.AllowedOriginsRaw,
-		"log_dir":        logHub.logDir,
-		"log_path":       logHub.logPath,
-		"log_dir_source": logHub.logDirSource,
-		"log_level":      logHub.logLevelName(),
-		"pid":            os.Getpid(),
-		"goos":           runtime.GOOS,
+		"log_dir":         logHub.logDir,
+		"log_path":        logHub.logPath,
+		"log_dir_source":  logHub.logDirSource,
+		"log_level":       logHub.logLevelName(),
+		"pid":             os.Getpid(),
+		"goos":            runtime.GOOS,
 	})
 	shutdownTelemetry, telemetryErr := initTelemetry(context.Background())
 	if telemetryErr != nil {
